chore: Extract Json preview into separate component
